tangle: Handle 'e' and 'E' consistently in real constants
In module "Special code to finish real constants" the branch to handle
a '+' or '-' sign updates cur_char with get_next(). However the code
that converts an 'e' to an 'E' is a separate branch that isn't reached
from here, so an 'e' will cause the loop to break, while an 'E' will
continue it.

Even though multiple E's in a number is invalid Pascal, as WEB goes to
some length to treat 'e' and 'E' as the same it makes sense to fix this
difference in behaviour.