weave: Fix zero-length identifier support
An empty identifier can be produced by delimited commands such as cross
references:

    @ @:@>

Two parts of TANGLE expect non-empty identifiers:

1. The hash in id_lookup() initialises the hash value with the first
   character. This will be the '@' in the '@>' sequence for an empty
   identifier produced by a cross reference.

2. Phase 3 initialises the sorting buckets by reading the first
   character of each identifier. This is unstable for empty identifiers;
   the identifier that comes next (after accounting for buffer striping)
   shares a byte_ptr value, so its first character determines the
   sorting order.

As an empty identifier is strange and won't render visibly in the output
anyway, add checks when scanning for cross references that harmlessly
drop these cases.

---

Alternatively, the two cases above can be fixed to handle empty
identifiers. The hash would start at 0 and iterate all the characters,
and phase 3 can treat the first character of an empty identifier as '\0'
or something.